fix: added constants, type, init to add event listeners
