+++ src/controller/playerController.ts
@@ -1,8 +1,9 @@
-import { CONTROLS } from '../constants/constants';
+import { CONTROLS, ERROR } from '../constants/constants';
 import { Player, Position } from '../model/player';
 import { Utility } from '../utility';
 import { BoardView } from '../views/BoardView';
 
+type GridRowAndCol = { row: number; column: number };
 export class PlayerController {
   view: BoardView;
   totalPlayers: number = 2;
@@ -10,12 +11,35 @@ export class PlayerController {
   players: Array<Player> = [];
   currentGrid!: Array<number[]>;
   utility: Utility;
-  rowAndCol!: { row: number; column: number };
+  rowAndCol!: GridRowAndCol;
   constructor(view: BoardView) {
+    this.init();
     this.view = view;
     this.utility = new Utility();
   }
-
+  init() {
+    const startBtn = document.getElementById('start');
+    if (startBtn) {
+      startBtn.addEventListener('click', () => {
+        this.handleStart();
+      });
+    }
+    const numberOfPlayers: HTMLSelectElement | null =
+      document.querySelector('#players');
+    if (numberOfPlayers) {
+      numberOfPlayers.addEventListener('change', (e: Event) => {
+        this.totalNoOfPlayers(e);
+      });
+    }
+    const difficultySelection = document.querySelector(
+      '.difficulty-selection',
+    ) as HTMLInputElement;
+    if (difficultySelection) {
+      difficultySelection.addEventListener('click', (e: Event) => {
+        this.selectGameDifficulty(e);
+      });
+    }
+  }
   createPlayers(totalPlayers: number) {
     for (let i = 0; i < totalPlayers; i++) {
       this.players.push(new Player(i, this.playerPosition[i], 0, i === 0));
@@ -28,8 +52,10 @@ export class PlayerController {
     this.view.displayGame(this.currentGrid, this.players);
   }
   handleController(e: Event) {
-    const currentPlayer = this.players.find((player) => player.turn)!;
-    this.handleOperations(e, currentPlayer);
+    const currentPlayer = this.players.find((player) => player.turn);
+    if (currentPlayer) {
+      this.handleOperations(e, currentPlayer);
+    }
   }
   handleOperations(e: Event, player: Player) {
     if (e.target instanceof HTMLElement) {
@@ -37,7 +63,7 @@ export class PlayerController {
       switch (typeOfControl) {
         case CONTROLS.UP: {
           if (player.position.x === 0) {
-            alert("You can't go up now!");
+            alert(ERROR.UP);
           } else {
             player.position.x -= 1;
             this.updateScoreAndGrid(this.currentGrid, player);
@@ -47,7 +73,7 @@ export class PlayerController {
         }
         case CONTROLS.DOWN: {
           if (player.position.x === this.rowAndCol.row - 1) {
-            alert("You can't go down now!");
+            alert(ERROR.DOWN);
           } else {
             player.position.x += 1;
             this.updateScoreAndGrid(this.currentGrid, player);
@@ -57,7 +83,7 @@ export class PlayerController {
         }
         case CONTROLS.LEFT: {
           if (player.position.y === 0) {
-            alert("You can't go left now!");
+            alert(ERROR.LEFT);
           } else {
             player.position.y -= 1;
             this.updateScoreAndGrid(this.currentGrid, player);
@@ -67,7 +93,7 @@ export class PlayerController {
         }
         case CONTROLS.RIGHT: {
           if (player.position.y === this.rowAndCol.column - 1) {
-            alert("You can't go right now!");
+            alert(ERROR.RIGHT);
           } else {
             player.position.y += 1;
             this.updateScoreAndGrid(this.currentGrid, player);
@@ -78,7 +104,7 @@ export class PlayerController {
       }
     }
   }
-  getGrid(arrObj: { row: number; column: number }) {
+  getGrid(arrObj: GridRowAndCol) {
     const coinGrid = this.utility.generateGridCoin(arrObj);
     this.playerPosition = this.utility.genrateRandom(this.totalPlayers, arrObj);
     return this.utility.clearPosition(this.playerPosition, coinGrid);
